backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlmodel>=0.0.18
python-multipart>=0.0.9
google-cloud-vision>=3.7.0
httpx>=0.27.0
python-dotenv>=1.0.0
aiofiles>=23.2.0
Pillow>=10.3.0
